- Fixes - i18n - Save/load selected categories to/from local storage - Lazy load the component
