+++ src/slices/explorer/types.ts
@@ -84,13 +84,11 @@ export type State = {
         msg: string;
         closeAfter?: number;
     } | null;
-    highlightSetting:
-        | undefined
-        | {
-              route: string;
-              accordion?: string;
-              field: string;
-          };
+    highlightSetting: null | {
+        route: string;
+        accordion?: string;
+        field: string;
+    };
 };
 
 export type PrimaryMenuConfigType = State["primaryMenu"];
